requirements.txt
pqcrypto>=0.4.0
psutil>=5.9.0
pyyaml>=6.0
pandas>=2.0.0
matplotlib>=3.7.0
fpdf2>=2.7.0
numpy>=1.24.0
scipy>=1.10.0
# Optional for comparison
# dilithium-py
